+++ km_api/know_me/serializers.py
@@ -77,10 +77,23 @@ class KMUserDetailSerializer(KMUserListSerializer):
         fields = KMUserListSerializer.Meta.fields + ('permissions', 'profiles')
 
 
+class KMUserAccessorAcceptSerializer(serializers.ModelSerializer):
+    """
+    Serializer for accepting an accessor.
+    """
+
+    class Meta:
+        fields = tuple()
+        model = models.KMUserAccessor
+
+
 class KMUserAccessorSerializer(serializers.HyperlinkedModelSerializer):
     """
     Serializer for ``KMUserAccessor`` instances.
     """
+    accept_url = serializers.HyperlinkedIdentityField(
+        view_name='know-me:accessor-accept')
+    permissions = DRYPermissionsField(additional_actions=['accept'])
     url = serializers.HyperlinkedIdentityField(
         view_name='know-me:accessor-detail')
     user_with_access = UserInfoSerializer(read_only=True)
@@ -91,12 +104,15 @@ class KMUserAccessorSerializer(serializers.HyperlinkedModelSerializer):
             'url',
             'created_at',
             'updated_at',
+            'accept_url',
             'email',
             'is_accepted',
             'is_admin',
             'km_user_id',
+            'permissions',
             'user_with_access')
         model = models.KMUserAccessor
+        read_only_fields = ('is_accepted',)
 
     def create(self, validated_data):
         """
@@ -112,37 +128,7 @@ class KMUserAccessorSerializer(serializers.HyperlinkedModelSerializer):
         km_user = validated_data.pop('km_user')
         email = validated_data.pop('email')
 
-        # Remove 'is_accepted' from the arguments so it's not passed to the
-        # 'share' method.
-        validated_data.pop('is_accepted', None)
-
         return km_user.share(email, **validated_data)
-
-    def validate(self, data):
-        """
-        Validate the data provided to the serializer as a whole.
-
-        Args:
-            data:
-                The data passed to the serializer, after each field has
-                been validated on its own.
-
-        Returns:
-            The validated data.
-        """
-        email = data.get('email')
-        km_user = data.get('km_user')
-
-        # The rest of the validation is only applicable when an email
-        # and Know Me user are provided.
-        if not all([email, km_user]):
-            return data
-
-        if km_user.km_user_accessors.filter(email=email).exists():
-            raise serializers.ValidationError(
-                _("%s has already been granted access to your account."))
-
-        return data
 
     def validate_email(self, email):
         """
@@ -166,35 +152,3 @@ class KMUserAccessorSerializer(serializers.HyperlinkedModelSerializer):
                 _('The email of an existing share may not be changed.'))
 
         return email
-
-    def validate_is_accepted(self, is_accepted):
-        """
-        Validate the provided 'is_accepted' value.
-
-        Args:
-            accepted (boolean):
-                A boolean indicating if the accessor has been accepted.
-
-        Returns:
-            boolean:
-                The validated 'is_accepted' value.
-
-        Raises:
-            serializers.ValidationError:
-                If the user doesn't have permission to change the
-                'is_accepted' attribute.
-        """
-        accessor = self.instance
-
-        if not accessor and is_accepted:
-            raise serializers.ValidationError(
-                _("An accessor can't be marked as accepted until after it has "
-                  "been created."))
-        elif (accessor
-                and accessor.is_accepted != is_accepted
-                and accessor.user_with_access != self.context['request'].user):
-            raise serializers.ValidationError(
-                _("Only the user granted access by the accessor may accept "
-                  "the accessor."))
-
-        return is_accepted