+++ km_api/know_me/urls.py
@@ -24,6 +24,11 @@ urlpatterns = [
         name='accessor-detail'),
 
     url(
+        r'^accessors/(?P<pk>[0-9]+)/accept/$',
+        views.AccessorAcceptView.as_view(),
+        name='accessor-accept'),
+
+    url(
         r'^config/$',
         views.ConfigDetailView.as_view(),
         name='config-detail'),